Rewrote HTML parser to convert malformed markup into a valid parse tree.
Jason and Kunal have a requirement that they be able to handle
malformed markup and are currently reimplementing event handler
extraction and other bits of HtmlPluginCompiler so that they can
hand us just javascript.

This change is meant to get us to the point where we can take
malformed markup as input and coerce it to the same parse tree
a browser would see.

It will also make it easier security-wise.  HTML doesn't guarantee
that if an id only shows up on one node, only one node in the
resulting DOM will have that id.  We can now coerce to XHTML which
has that property.




I suggest you review in this order:
 # New parser changes
 java/.../parser/html/DomParser.java
 java/.../parser/html/OpenElementStack.java
 java/.../parser/html/XmlElementStack.java
 javatests/.../parser/html/DomParserTest.java
 java/.../parser/html/Html5ElementStack.java

 # Changes to rendering of HTML doms
 java/.../parser/DomTree.java
 java/.../lexer/HtmlLexer.java
 java/.../lexer/HtmlTextEscapingMode.java

 # Everything else



This adds a third-party dependency on libhtmlparser available
at http://about.validator.nu/htmlparser/ which is available
under the MPL.  I downloaded the most recent version, 1.5.1.
It implements HTML5's (
http://www.whatwg.org/specs/web-apps/current-work/ )
parsing rules which attempt to codify the behavior of existing browsers
when parsing HTML.



Changes:
- Updated HtmlLexer to properly distinguish RCDATA and CDATA.
 RCDATA is the content type for <title> and <textarea> tags
 which interpret special characters literally, but which
 treat entities as entities.  Fixed lexer to not recognize
 CDATA sections when parsing HTML.

- Modified DomParser to canonicalize element and attribute names.
 Existing code tries to normalize case before comparing element
 names to whitelists, but this is error-prone (and incorrect for
 XHTML), so I now canonicalize at construction time.

- Rewrote DomTree.render to properly render unary HTML tags.
 Before we output <br></br> for a single <br> tag, but browsers
 actually treat a </br> token as a <br> token.

- Defined OpenElementStack, an interface used by DomParser to
 build the tree and implemented a trivial one for XML parsing.

- Defined Html5ElementStack which bridges OpenElementStack to
 libhtmlparser.  This uses our lexer, and parse tree implementation
 and preserves file position info for both nodes and error messages.

- fixed bug: NPE on link tags in gadget specs that specify the
 media attribute but not type.




git-svn-id: http://google-caja.googlecode.com/svn/trunk@362 360991c0-9f3b-0410-ba87-8b362a00edf4