Added flags to permissive.js, to caja.js's call to the logging function, and to log-to-console's implementation of the logging function.
Notice of non-upward compatible change: permissive.js now loads in a
disabled state and must be explicitly enabled. This is needed to
support a checkbox in the open-social sample container. Once enabled,
disabling it does not undo the damage. The page should be reloaded
instead. 

The optional flag added to the logging function tells it that normal
control flow is about to be terminated by a thrown exception, to give
it the opportunity to terminate control flow by other
means. log-to-console makes use of this in the Firebug environment in
order to invoke a non-existent method. This causes Firebug to wake up
with a live stack trace.




git-svn-id: http://google-caja.googlecode.com/svn/trunk@426 360991c0-9f3b-0410-ba87-8b362a00edf4