*mikesamuel/lex-NaN-Infinity-as-words-11-Mar-2008@725 | mikesamuel | 2008-03-11 20:31:00 -0800 (Tue, 11 Mar 2008)
Description:

To address bug 67, we need to treat NaN as a regular identifier.

This changes NaN, and Infinity which had been lexed as floating point
tokens, and now treats them as non-reserved words.

Also updates the renderer to render numeric literals properly.  If we
fold constants and get a numeric value of NaN or Infinity, we will now
render that parse tree in a manner that is independent of environment.






git-svn-id: http://google-caja.googlecode.com/svn/trunk@738 360991c0-9f3b-0410-ba87-8b362a00edf4