*mikesamuel/caja-candelete-strawman@961 | mikesamuel | 2008-04-01 19:59:56 -0800 (Tue, 01 Apr 2008)
Description:

A cut at {delete,canDelete}{Pub,Prop} in caja.js.

I don't know whether this is the right thing to do, because I don't know how deletion works with whitelisting, or whether a field that was settable and has been deleted should still be settable.

(1) Does not allow deletion on frozen objects, or deletion of foo__ fields.
(2) Allows deletion of o.foo and o.foo_ on JSON objects
(3) Allows deletion of o.foo but not o.foo_ on non-JSON objects
    unless ___.allowDelete(o, 'foo')
(4) Allows deletion of this.foo and this.foo_ from methods.







git-svn-id: http://google-caja.googlecode.com/svn/trunk@963 360991c0-9f3b-0410-ba87-8b362a00edf4