*mikesamuel/css-rewriter-errors-to-warnings@966 | mikesamuel | 2008-04-02 12:33:06 -0800 (Wed, 02 Apr 2008)
Description:

Add a mode switch to CssRewriter and CssValidator that lets the
messages issued when a node is marked invalid be treated as warnings
as long as they're removed later.

When we believe the styles come from noisy HTML, we treat invalid
nodes as warnings and simply remove them from the tree.





git-svn-id: http://google-caja.googlecode.com/svn/trunk@967 360991c0-9f3b-0410-ba87-8b362a00edf4