*mikesamuel/lexer-fixes-and-minimal-renderer@1135 | mikesamuel | 2008-04-18 12:18:22 -0800 (Fri, 18 Apr 2008)
Description:

Fix bug: sequences like i---j where a sequence of characters s is a
punctuation and a non-zero length prefix of s is a suffix of another
punctuation string could result in JsLexer generating a punctuation
token that is not in the grammar.

This caused problems with lexing minified input.

As part of investigating this I implemented a minimzing JS renderer so
I could use the pretty printer random test case generator to find
other holes in the lexer.

- fixed bug in InputElementSplitter
- updated Punctuation to take into account changes to the ES4 spec.
  We want ES4 to correctly lex Caja output even though we only need
  to parse ES3.  Added .. operator and removed the now defunct ^^.
- Removed code from InputElementJoiner that is now unused since there
  is no longer any suffix of a punctuation string that is a prefix of
  a punctuation string that is not also itself a punctuation string.
- Implemented JsMinimalPrinter which should allow us to do a minified
  output mode.



git-svn-id: http://google-caja.googlecode.com/svn/trunk@1159 360991c0-9f3b-0410-ba87-8b362a00edf4