*mikesamuel/fix-bug-94-Array-sort@1130 | mikesamuel | 2008-04-17 11:25:34 -0800 (Thu, 17 Apr 2008)
Description:

Fix bug 94: Array.sort allows confused deputy attacks

caja.js calls allowMutator on Array.prototype.sort.

Array.sort invokes the comparator with this bound to the global scope.

Nothing forces the comparator to be called as a simple function, so a
function received on an object passed in can be invoked with this bound to
the global scope.


Consider the object passed in
  outers.deputy = {
      f: function (a) { return this.alert(a); },
      alert: function (s) { /* something innocuous */ }
  };
and the caja
  ['hello world', 'hello world'].sort(deputy.f);


<script type="text/javascript">
['alert("hello")', 'alert("world")'].sort(eval)
</script>








git-svn-id: http://google-caja.googlecode.com/svn/trunk@1186 360991c0-9f3b-0410-ba87-8b362a00edf4