*mikesamuel/minimal-printer-fix-27-Apr-2008@1199 | mikesamuel | 2008-04-27 22:56:54 -0800 (Sun, 27 Apr 2008)
Description:

Fix bug: minimal printer was confused since the Trie with "<!--" and
"]]>" added had non-terminal nodes which means to be strictly correct,
we need to keep one-trie per space not inserted.  This is probably not
exploitable, but fixed to make all Trie nodes terminal.

Also made sure that newlines were not inserted where they are illegal
in a "restricted production" as specified in Section 7.9.1 of ES262.




git-svn-id: http://google-caja.googlecode.com/svn/trunk@1213 360991c0-9f3b-0410-ba87-8b362a00edf4