*mikesamuel/ctors-are-not-simple@1668 | mikesamuel | 2008-06-04 16:02:27 -0800 (Wed, 04 Jun 2008)
Description:

Issue 217: Prevent functions marked as simple from being marked as constructors as well.

asCtor allows simple functions and constructors to appear after new.

But caja.def calls ctor, which used to prohibit simple functions in
the first position.  (The check was removed in
metaweta/attached-methods.)  This leaves the possibility that a
function might be marked both as a contructor and a simple function,
which messes with the function species table:
http://code.google.com/p/google-caja/wiki/FunctionSpecies


- Add tests to make sure that simple functions can't be modified after
  first call.
- Add check to make sure that exophoric functions can't be constructors.
- Split the code that attaches the base class to a constructor out of
  ctor, and call it directly from caja.def if the input function is
  simple.




git-svn-id: http://google-caja.googlecode.com/svn/trunk@1673 360991c0-9f3b-0410-ba87-8b362a00edf4