*mikesamuel/fix-client-side-css-validator@2094 | mikesamuel | 2008-08-05 11:33:13 -0800 (Tue, 05 Aug 2008)
Description:

A bug was causing us to incorrectly reject CSS <lengths> without a sign.

"left: 10px" was being rejected, whereas "left: +10px" was fine.

I added unittests, but a Rhino bug caused them to fail.  I filed that as
    https://bugzilla.mozilla.org/show_bug.cgi?id=448443
but the workaround was to remove the extra parentheses.

I also update rhino/js.jar to version 1.7 to see if that bug was fixed
in a new version, but it wasn't.



git-svn-id: http://google-caja.googlecode.com/svn/trunk@2142 360991c0-9f3b-0410-ba87-8b362a00edf4