*mikesamuel/bug-741@2432 | mikesamuel | 2008-09-29 01:59:43 -0800 (Mon, 29 Sep 2008)
Description:

bug 741: java.lang.NullPointerException when cajoling a specific CSS rule.
  index.html:
    <link rel="Stylesheet" media="screen" href="testsuite.css" />
  testsuite.css:
    div#show-tests * { display: none; }

It had nothing to do with the CSS rule, but with the link tag that has
a media attribute but no type attribute.

Added tests and fixed RewriteHtmlStage to use the media attribute file
position for generate @media blocks instead of the type attribute's
file position.



git-svn-id: http://google-caja.googlecode.com/svn/trunk@2476 360991c0-9f3b-0410-ba87-8b362a00edf4