*mikesamuel/bug-815@2712 | mikesamuel | 2008-10-20 17:28:38 -0800 (Mon, 20 Oct 2008)
Description:

Fixes bugs 815 and 845

Issue 815:  	 Inline style= attributes don't work in ie
    This is similar to Issue 636 , except its in html-emitter and not
    domita.

    <div style="border: 1px solid red;">foo</div>

    How about moving this setAttribute logic to bridal and sharing it
    between html-emitter and domita?


Issue 845:  	 IE can't set innerHTML of table element

    html-emitter's ih method fails with an unknown error when used to
    inject the content of a table element.

    This fails on IE in squarefree shell and in domita_test.html:
    var table = document.createElement('table');
    table.innerHTML = '\n   <tbody><tr>\n   <td><strong>'
        + 'Click me if this is red</strong></td>\n   </tr>\n  </tbody>';
    Error: Unknown runtime error


Moved setAttribute code, that switches on the attribName possibly
using style.cssText= instead, from DOMita to bridal, and changed
html-emitter to use it.

Testing that flushed out an IE bug whereby setting the innerHTML of a
table element causes an exception on IE.  To get domita_test.html to
run green on IE, I fixed that in DomProcessingEvents, and put together
a file (linked from bug 845) that checks that the innerHTML of various
elements are settable.




git-svn-id: http://google-caja.googlecode.com/svn/trunk@2713 360991c0-9f3b-0410-ba87-8b362a00edf4