*mikesamuel/myvn-eclipse-lib-overlap@2639 | mikesamuel | 2008-10-17 11:14:46 -0800 (Fri, 17 Oct 2008)
Description:

Changes (myvn eclipse) to generate a .classpath that does not include
all the .class and .jar files from ant-lib, but only those that are
built by ANT and not by eclipse.

This fixes C-f in eclipse so that one doesn't see multiple versions of
resources, and see .class files when opening for a java file.





git-svn-id: http://google-caja.googlecode.com/svn/trunk@2725 360991c0-9f3b-0410-ba87-8b362a00edf4