Fixes 844, 843, 834, 847, 502, 854.
844: privilege escalation:

reifyIfXo4a(), when given a tamed exophoric function (which therefore
may not be first-class), returns a malfunction -- a record with call,
apply, and bind methods that emulate and wrap the safe behavior of the
original function. However, these methods forgot to check whether
their first argument was null or undefined. By passing these
unmodified to the original function, an attacker can cause the
original function to operate instead on the global object. By
replacing such bad this-values with cajita.USELESS, we plug this leak.

834: More broken callbacks:

Several parts of the Cajita runtime as well as several tamed parts of
the ES3 API take callbacks which were assumed to be
simple-functions. However, to accommodate Valija's disfunctions and
now Cajita's malfunctions, these callbacks accept as well a
non-function object with a callable apply() method. The callback is
always then performed by invoking the callback argument's apply()
method.

DisfunctionPrototype is now initialized with generic disfunctions for
call, apply, and bind. We no longer create per-disfunction non-generic
bind methods, though we still do for call and apply.

843: Testbed stopped displaying outcomes:

The new/old rewrite of an entire html file as a simple module, with a
single call to loadModule, returned only the result of the last script
block in the file, as opposed to the last outcome. We now initialize
the moduleResult___ to ___.NO_RESULT, so that we can avoid overwriting
a report of an abrupt outcome. This allows us to distinguish the
situation where the last script block threw an error from the
situation where the last script block returned an accessible
value. (For this to work, ___.NO_RESULT should never be made
accessible to untrusted code.)

The testbed's display logic now reports the outcome of the outer
module, not the inner module, since the outer module is now the only
thing we've got. It also specially reports access to the actual global
object, as that represents and especially interesting security
breach. (This ability helped develop the fix for #844.)

847: '\w' that should have been '\\w'

Trivial fix. Interactively tested the result on Felix's example in the
testbed. 

502: derive() and ctor()'s opt_Sup argument's purpose

Bug was actually just a documentation problem. Also moved an if, as
just suggested by David-Sarah, in order to make the code clearer.

854: Valija's read() ($v.r()) was giving a strange error if asked to
read a property from null or undefined. It now checks for this case
explicitly, and throws an informative TypeError.

Other changes

* cajita-debugmode.js was replacing cajita.log with an instrumented
  one, but forgot to mark it as a simple frozen function.

* cajita.BREAK and cajita.USELESS are now statically permitted as
  CanRead rather than CanCall. This does not yet have any affect.


git-svn-id: http://google-caja.googlecode.com/svn/trunk@2767 360991c0-9f3b-0410-ba87-8b362a00edf4