*mikesamuel/better-pretty-printing@2948 | mikesamuel | 2008-11-11 20:12:02 -0800 (Tue, 11 Nov 2008)
Description:

Split out the bits of JavaScript rendering code relating to deciding
whether spaces or newlines are required/allowed between tokens.

Ihab pointed out that the side-by-side renderer is splitting lines in
ways that allow for semicolons to be inserted after return and throw.

This consolidates the code related to that.  In another CL I will
rewrite existing renderers to use this consistently, and in a third, I
will rewrite JsPrettyPrinter to do better formatting.  This CL is
simple and semantics preserving.  The second will be a more
substantial change but semantics preserving.  The third will not be
semantics preserving.





git-svn-id: http://google-caja.googlecode.com/svn/trunk@2977 360991c0-9f3b-0410-ba87-8b362a00edf4