Issue 10242  http://codereview.appspot.com/10242 Fix DOMita tests and test clearTimeout & clearInterval.
Implements getElementsByClassName as discussed in bug 933, since the
lack of an implementation was causing testEditabilityInNodeLists to
fail.

Moved some of the tests at the top lower down so that they don't make
it hard to see the interactive tests.

Added tests for clearTimeout & clearInterval since it seems that
clearTimeout and clearInterval are failing in prototype tests though
this could be due to race conditions.

R=metaweta@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3104 360991c0-9f3b-0410-ba87-8b362a00edf4