Issue 9662  http://codereview.appspot.com/9662 Reimplement pretty printer and minimizer on top of BufferedRenderer
Review at http://codereview.appspot.com/9662/show

This change is the second part of the pretty printer rewrite.

It attempts to handle several bugs that occur when comments are
allowed in rendered text as in the side-by-side renderer.

In the program
(function () { a: for (;;) { for (;;) { break/*
*/a } return 1 } return 2 })()
the comment may be lexically significant.

This change fixes that by recognizing comments that fall inside
restricted productions.

R=ihab.awad


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3113 360991c0-9f3b-0410-ba87-8b362a00edf4