Issue 10444  http://codereview.appspot.com/10444 Change ParserBase to compile under JDK1.5.
ParserBase has a compile time dependency on java.text.Normalizer,
though some tricks are done to avoid a runtime dependency.

Tobie needs to compile on an old version of MacOS X, so this
uses nasty reflection to get rid of the compile time dependency.

It also adds comments around the fallback code that David-Sarah
dug up, and expands to all of ISO-8859-1.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3115 360991c0-9f3b-0410-ba87-8b362a00edf4