Issue 12243  http://codereview.appspot.com/12243 Fix problems with testbed in IE
IE was failing for two reasons.

(1) The two CSS height properties were causing the result and error
message boxes to be tiny making it impossible to read content.
Changed one height to min-height which works on FF and Safari and does
not break IE 6.

(2) Something in domita is calling console.log which adds a log output
element to document.body.  This is invoked from script in the head
before IE creates a document.body.  Moved script loading into the
body.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3214 360991c0-9f3b-0410-ba87-8b362a00edf4