Issue 14086  http://codereview.appspot.com/14086 Fix HTML renderer to prevent rendering of unrenderable CDATA content..
The XHTML document
  <xmp> &lt/xmp&gt; </xmp>
is not translatable to an HTML document since the close equivalent
  <xmp> </xmp> </xmp>
has a different meaning.

This change fixes this to cause rendering CDATA content in HTML
rendering mode that contains a sequence of characters that would close
the CDATA element to fail with an IllegalStateException.

This is not a breach vector since no CDATA content elements appear on
our schema whitelist, hence no need for responsible disclosure, but it
is a correctness and maintainability issue.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3273 360991c0-9f3b-0410-ba87-8b362a00edf4